+++ frontend/src/components/Question.vue
@@ -1,0 +1,56 @@
+<script setup lang="ts">
+import BetaversionLabel from '@/components/betaversion-label.vue'
+import Sources from '@/components/Sources.vue'
+import { Answer } from '@/models/DecisionTree'
+
+interface Props {
+  id: string
+  question: string
+  sources: { source: string; url: string | undefined; }[] | undefined
+  answers: Array<Answer>
+}
+
+defineProps<Props>()
+defineEmits(['answered'])
+</script>
+
+<template>
+  <legend>
+    <div class="flex items-center">
+      <!--    TODO: Replace question with categories-->
+      <h1 class="utrecht-heading-1">Vraag:</h1>
+      <span
+        class="inline-flex items-center rounded-md bg-blue-50 px-2 py-1 text-xs font-medium text-blue-700 ring-1 ring-inset ring-blue-700/10"
+      >{{ id }}</span
+      >
+      <BetaversionLabel />
+    </div>
+    <Sources :sources="sources" />
+
+    <!-- Question section -->
+    <div class="mt-6 space-y-6">
+      <p style="white-space: pre-line" class="inline-flex utrecht-paragraph">
+        <span v-html="question"></span>
+        <slot />
+      </p>
+    </div>
+
+    <!-- Answers section -->
+    <div class="mt-6 space-y-6">
+      <div v-for="(answer, index) in answers" :key="index" class="flex items-center gap-x-3">
+        <input
+          :key="id + index.toString()"
+          aria-roledescription="radio button"
+          @click="$emit('answered', answer)"
+          :value="id + index.toString()"
+          :id="index.toString()"
+          type="radio"
+          class="h-4 w-4 border-gray-300 text-indigo-600 focus:ring-indigo-600"
+        />
+        <label :for="index.toString()" class="block text-sm font-medium leading-6 text-gray-900">{{
+            answer.answer
+          }}</label>
+      </div>
+    </div>
+  </legend>
+</template>
