+++ frontend/src/components/betaversion-label.vue
@@ -1,0 +1,13 @@
+<template>
+  <div
+    class="flex justify-center rounded-md rvo-alert rvo-alert--padding-xs rvo-alert--warning">
+            <span
+              class="utrecht-icon rvo-icon rvo-icon-waarschuwing rvo-icon--xl rvo-status-icon-waarschuwing"
+              role="img"
+              aria-label="Waarschuwing"
+            ></span>
+    <div class="flex utrecht-alert--warning utrecht-paragraph">
+      bètaversie
+    </div>
+  </div>
+</template>
